requirements.txt
asgiref==3.8.1
Django==5.0.7
django-db-connection-pool==1.2.5
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
Markdown==3.6
mysql-connector-python==9.0.0
mysqlclient==2.2.4
PyJWT==2.8.0
SQLAlchemy==2.0.31
sqlparams==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
